+++ packages/api/src/platforms/vtex/resolvers/organization.ts
@@ -1,9 +1,9 @@
 import type { Resolver } from '..'
-import { PromiseType } from '../../../typings'
-import { StoreOffer } from './offer'
+import type { PromiseType } from '../../../typings'
+import type { StoreOffer } from './offer'
 
 export type Root = PromiseType<ReturnType<typeof StoreOffer.seller>>
 
-export const StoreOrganization: Record<string,  Resolver<Root>> = {
+export const StoreOrganization: Record<string, Resolver<Root>> = {
   identifier: () => '',
 }