+++ packages/sdk/types/index.d.ts
@@ -1,5 +1,5 @@
 declare global {
-  var faststore_sdk_stores: Map<string, Store> 
+  var faststore_sdk_stores: Map<string, Store>
 }
 
 export {}